🎨The function: Adding a customer, externally and using it instead of what was there until now, is written in two places.
